Pressing left ctrl while bags are open will now flash the cheapest grey item in your bags
so that you can easily find the cheapest item in your bags to delete if your bags are full and not close to a vendor. Works for default bags and OneBag3 addon only, if you are using some other bag addon and the feature does not work just leave a comment on what bag addon you are using so that I can add support for it.